door-key-open-detection-with-tf/app/requirements.txt
tensorflow==2.9.0
pillow==9.0.1
protobuf==3.19.4